+++ pkg/traceql/engine_metrics_test.go
@@ -378,6 +378,45 @@ func TestCompileMetricsQueryRange(t *testing.T) {
 	}
 }
 
+func TestCompileMetricsQueryRangeExemplarsHint(t *testing.T) {
+	defaultExempalars := 5
+
+	tcs := []struct {
+		q             string
+		expectedCount int
+	}{
+		{
+			q:             "{} | rate() with(exemplars=10)",
+			expectedCount: 10,
+		},
+		{
+			q:             "{} | rate() with(exemplars=false)",
+			expectedCount: 0,
+		},
+		{
+			q:             "{} | rate() with(exemplars=true)",
+			expectedCount: defaultExempalars,
+		},
+		{
+			q:             "{} | rate()",
+			expectedCount: defaultExempalars,
+		},
+	}
+
+	for _, tc := range tcs {
+		eval, err := NewEngine().CompileMetricsQueryRange(&tempopb.QueryRangeRequest{
+			Query: tc.q,
+			Start: 1,
+			End:   2,
+			Step:  1,
+		}, 5, 0, false)
+
+		require.NoError(t, err)
+		require.NotNil(t, eval)
+		require.Equal(t, tc.expectedCount, eval.maxExemplars)
+	}
+}
+
 func TestCompileMetricsQueryRangeFetchSpansRequest(t *testing.T) {
 	tc := map[string]struct {
 		q           string